+++ docs/changelog.md
@@ -5,6 +5,12 @@
 ## Changelog
 
 
+### v0.2.3
+
+- Refactor: use configparser.
+
+Compatibility: novelyst v4.36 API
+
 ### v0.2.2
 
 - Modify message (#1)